+++ src/CoffeeMachine.Domain/Models/Purchase.cs
@@ -2,15 +2,15 @@
 
 public class Purchase
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     
     public string Status { get; set; }
     public DateTime Date { get; set; }
     
-    public int IdCoffee { get; set; }
+    public long IdCoffee { get; set; }
     public Coffee Coffee {get; set;}
     
-    public int IdMachine { get; set; }
+    public long IdMachine { get; set; }
     public Machine Machine { get; set; }
 
     public List<Transaction> Transactions { get; set; } = new List<Transaction>();
