+++ src/CoffeeMachine.Domain/Models/Transaction.cs
@@ -2,14 +2,14 @@
 
 public class Transaction
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public bool Type { get; set; } // true - оплата, false - сдача
     public int CountBanknotes { get; set; } // кол-во одинаковых банкнот
     
-    public int IdBanknote { get; set; }
+    public long IdBanknote { get; set; }
     public Banknote Banknote { get; set; }
     
-    public int IdPurchase { get; set; }
+    public long IdPurchase { get; set; }
     public Purchase Purchase { get; set; }
 
 }
